+++ lib/events/export.dart
@@ -1,3 +1,5 @@
 export 'events.dart';
 export 'subscription_events.dart';
 export 'auto_catching_type.dart';
+export 'on_send_success_callback.dart';
+export 'on_send_failed_callback.dart';