+++ lib/events/on_send_failed_callback.dart
@@ -1,0 +1,3 @@
+import '../network/http_response.dart';
+
+typedef OnSendFailedCallback = void Function(HttpResponse status);