+++ lib/native/native_base.dart
@@ -1,11 +1,12 @@
-import '../utils/try_cast.dart';
-import '../utils/uuid.dart';
+import '../utils/export.dart';
 import 'affise_api_method.dart';
 import 'native_base_platform.dart';
 
 class NativeBase extends NativeBasePlatform {
   static const _UUID = "callback_uuid";
+  static const _TAG = "callback_tag";
 
+  final Map<String, Map<String, dynamic>> _callbacksGroup = {};
   final Map<String, dynamic> _callbacksOnce = {};
   final Map<AffiseApiMethod, dynamic> _callbacks = {};
 
@@ -26,6 +27,16 @@ class NativeBase extends NativeBasePlatform {
     apiCall(api.apiName, apiData);
   }
 
+  void nativeCallbackGroup(AffiseApiMethod api, Map<String, dynamic> callbackGroup, [dynamic data]) {
+    String uuid = Uuid.generate();
+    Map<String, dynamic> apiData = {
+      api.apiName: data,
+      _UUID: uuid,
+    };
+    _callbacksGroup[uuid] = callbackGroup;
+    apiCall(api.apiName, apiData);
+  }
+
   void nativeCallback(AffiseApiMethod api, dynamic callback, [dynamic data]) {
     Map<String, dynamic> apiData = {
       api.apiName: data,
@@ -35,21 +46,33 @@ class NativeBase extends NativeBasePlatform {
   }
 
   @override
-  void apiCallback(String apiName, dynamic arguments) {
+  dynamic apiCallback(String apiName, dynamic arguments) {
     AffiseApiMethod? api = apiMethodFrom(apiName);
-    if (api == null) return;
+    if (api == null) return null;
     Map<Object?, Object?>? map = tryCast<Map<Object?, Object?>>(arguments);
     String? uuid = map?[_UUID]?.toString();
+    String? tag = map?[_TAG]?.toString();
     dynamic data = map?[apiName];
+
+    dynamic result;
+
     if (uuid == null || uuid.isEmpty) {
       dynamic callback = _callbacks[api];
-      handleCallback(api, callback, data);
+      result = handleCallback(api, callback, data, null);
+    } else if (tag != null && tag.isNotEmpty) {
+      dynamic callback = _callbacksGroup[uuid]?[tag];
+      result = handleCallback(api, callback, data, tag);
+      _callbacksGroup.remove(uuid);
     } else {
       dynamic callback = _callbacksOnce[uuid];
-      handleCallback(api, callback, data);
+      result = handleCallback(api, callback, data, null);
       _callbacksOnce.remove(uuid);
     }
+
+    return result;
   }
 
-  void handleCallback(AffiseApiMethod api, dynamic callback, dynamic data) {}
+  dynamic handleCallback(AffiseApiMethod api, dynamic callback, dynamic data, String? tag) {
+    return null;
+  }
 }